pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "liquor-marketing-agent"
version = "0.1.0"
description = "Two-agent marketing stack for liquor retail"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "python-dotenv>=1.0",
  "click>=8.1"
]

[project.optional-dependencies]
openai = ["openai>=1.40"]
dev = ["black>=24.0.0", "isort>=5.12.0", "flake8>=7.0.0", "pytest>=7.4.0"]
mail = ["requests>=2.31"]
sms = ["twilio>=9.0.0"]

[project.scripts]
liquor-subagent = "liquor_agent.subagent:main"
liquor-plan = "liquor_agent.orchestrator:main"
liquor-send = "liquor_agent.sender:main"

[tool.black]
line-length = 100
